Add negative test for test_svm size check
Isuue #875 clarified that CL_INVALID_BUFFER_SIZE should be returned
if clCreateBuffer is passed a pointer returned by clSVMAlloc as its host_ptr
and the size of the buffer exceeds the size of the SVM allocation.

Add a new negative test to ensure CL_INVALID_BUFFER_SIZE
is returned when the size of buffer exceeds the size of the SVM
allocation.

Fixes #1701